purge-logs.py: Do not try to remove symlink
2020-09-10 05:21:44,881 INFO  /var/www/logs/96/24796/12/check/rdoinfo-DLRN-check-centos8/c259e77/buildset/centos-rpm-master/repos/component/common/current : removing old logs
Traceback (most recent call last):
  File "/usr/local/bin/purge-logs.py", line 100, in <module>
    search_and_destroy(calculated_time, args.dry_run, root)
  File "/usr/local/bin/purge-logs.py", line 94, in search_and_destroy
    delete_dir(job_dir)
  File "/usr/local/bin/purge-logs.py", line 35, in delete_dir
    shutil.rmtree(dir_path)
  File "/usr/lib64/python3.6/shutil.py", line 496, in rmtree
    onerror(os.path.islink, path, sys.exc_info())
  File "/usr/lib64/python3.6/shutil.py", line 494, in rmtree
    raise OSError("Cannot call rmtree on a symbolic link")
OSError: Cannot call rmtree on a symbolic link

+++ ansible/roles/sf-logserver/files/purge-logs.py
@@ -99,6 +99,9 @@ def find_old_files(calculated_time, log_path):
 
 def search_and_destroy(calculated_time, dry_run, log_path):
     for job_dir in find_old_files(calculated_time, log_path):
+        if os.path.islink(job_dir):
+            log.debug("%s : skipping symlink", job_dir)
+            continue
         log.debug("%s : removing old logs", job_dir)
         if not dry_run and log_path != job_dir:
             delete_dir(job_dir)
